Add ability to pass args to cli API
Typically this function is called in a callable script, and therefore
reads in arguments through `sys.argv`. However, there are situations
(testing, debugging, etc) where one would like to emulate the callable
script programmatically. This commit makes this possible by adding an
`args` parameter to the `CLIInterfaceAPI.__call__` method. If no
args are passed, it will default to `sys.argv`.